requirements.txt
aiogram~=3.1.1
requests~=2.31.0
Pillow~=10.0.1